[u-boot] Specify serverip in tftpboot commands
The DHCP server used on the network returns a 'next server' value.
This is used by the 'dhcp' command in u-boot to set the 'serverip' variable.
This is causing issues for the PicoZed boards, as the next server value does not match the tftp server ip address that we require.

To work around this, we can explicitly provide the IP address in the tftpboot commands.

[ICSHWI-6818]